Factor out tinygraph url formatting to utils file
